FEATURE: Two new badges - First Emoji and First Mention
