+++ lib/cooked_post_processor.rb
@@ -28,6 +28,23 @@ class CookedPostProcessor
       post_process_oneboxes
       optimize_urls
       pull_hotlinked_images(bypass_bump)
+      create_firsts
+    end
+  end
+
+  def create_firsts
+    created = false
+
+    if @doc.css("img.emoji").size > 0
+      created |= UserFirst.create_for(@post.user_id, :used_emoji, @post.id)
+    end
+
+    if @doc.css("span.mention, a.mention").size > 0
+      created |= UserFirst.create_for(@post.user_id, :mentioned_user, @post.id)
+    end
+
+    if created
+      BadgeGranter.queue_badge_grant(Badge::Trigger::PostProcessed, user: @post.user)
     end
   end
